+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Impl `Eq` for `Config`;
 - `Config` builder;
+- Impl `FromStr` for `Config`;
 
 ## [5.2.0] - 2026-04-28